pa-workflow/requirements.txt
# Python dependencies
# Frameworks
fastapi
uvicorn
langgraph
pydantic-settings

# Auth
python-jose[cryptography]
passlib[bcrypt]

# Database
psycopg2-binary
pymongo
redis
sqlalchemy[asyncio]
asyncpg
motor
aioredis

# AWS
boto3

# Testing
pytest
httpx
